+++ src/swarms.c
@@ -20,6 +20,7 @@
 #include "party_menu.h"
 #include "list_menu.h"
 #include "overworld.h"
+#include "pokedex.h"
 #include "item.h"
 #include "constants/form_change_types.h"
 #include "constants/items.h"
@@ -29,7 +30,7 @@
 
 u8 swarmsActivated = 0;
 
-bool8 GetSwarmsDone(void)
+bool8 AreSwarmsDone(void)
 {
     return swarmsActivated == NUM_SWARMS_FLAGS;
 }
@@ -206,6 +207,112 @@ u16 GetSwarmSpeciesFlag(u16 species)
     }
 }
 
+u16 GetSpeciesFromSwarmFlag(u16 swarmFlag)
+{
+    switch (swarmFlag)
+    {
+    case FLAG_CATERPIE_SPECIES_SWARMING:
+        return SPECIES_CATERPIE;
+    case FLAG_WEEDLE_SPECIES_SWARMING:
+        return SPECIES_WEEDLE;
+    case FLAG_PIDGEY_SPECIES_SWARMING:
+        return SPECIES_PIDGEY;
+    case FLAG_RATTATA_SPECIES_SWARMING:
+        return SPECIES_RATTATA;
+    case FLAG_SPEAROW_SPECIES_SWARMING:
+        return SPECIES_SPEAROW;
+    case FLAG_EKANS_SPECIES_SWARMING:
+        return SPECIES_EKANS;
+    case FLAG_NIDORAN_SPECIES_SWARMING:
+        return SPECIES_NIDORAN_F;
+        // NIDORAN_M species is handled by the code that calls this function
+        // return SPECIES_NIDORAN_M;
+    case FLAG_CLEFAIRY_SPECIES_SWARMING:
+        return SPECIES_CLEFAIRY;
+    case FLAG_MEOWTH_SPECIES_SWARMING:
+        return SPECIES_MEOWTH;
+    case FLAG_MANKEY_SPECIES_SWARMING:
+        return SPECIES_MANKEY;
+    case FLAG_GROWLITHE_SPECIES_SWARMING:
+        return SPECIES_GROWLITHE;
+    case FLAG_POLIWAG_SPECIES_SWARMING:
+        return SPECIES_POLIWAG;
+    case FLAG_BELLSPROUT_SPECIES_SWARMING:
+        return SPECIES_BELLSPROUT;
+    case FLAG_PONYTA_SPECIES_SWARMING:
+        return SPECIES_PONYTA;
+    case FLAG_SLOWPOKE_SPECIES_SWARMING:
+        return SPECIES_SLOWPOKE;
+    case FLAG_SEEL_SPECIES_SWARMING:
+        return SPECIES_SEEL;
+    case FLAG_SHELLDER_SPECIES_SWARMING:
+        return SPECIES_SHELLDER;
+    case FLAG_GASTLY_SPECIES_SWARMING:
+        return SPECIES_GASTLY;
+    case FLAG_ONIX_SPECIES_SWARMING:
+        return SPECIES_ONIX;
+    case FLAG_DROWZEE_SPECIES_SWARMING:
+        return SPECIES_DROWZEE;
+    case FLAG_KRABBY_SPECIES_SWARMING:
+        return SPECIES_KRABBY;
+    case FLAG_EXEGGCUTE_SPECIES_SWARMING:
+        return SPECIES_EXEGGCUTE;
+    case FLAG_CUBONE_SPECIES_SWARMING:
+        return SPECIES_CUBONE;
+    case FLAG_LICKITUNG_SPECIES_SWARMING:
+        return SPECIES_LICKITUNG;
+    case FLAG_CHANSEY_SPECIES_SWARMING:
+        return SPECIES_CHANSEY;
+    case FLAG_TANGELA_SPECIES_SWARMING:
+        return SPECIES_TANGELA;
+    case FLAG_KANGASKHAN_SPECIES_SWARMING:
+        return SPECIES_KANGASKHAN;
+    case FLAG_SCYTHER_SPECIES_SWARMING:
+        return SPECIES_SCYTHER;
+    case FLAG_JYNX_SPECIES_SWARMING:
+        return SPECIES_JYNX;
+    case FLAG_ELECTABUZZ_SPECIES_SWARMING:
+        return SPECIES_ELECTABUZZ;
+    case FLAG_MAGMAR_SPECIES_SWARMING:
+        return SPECIES_MAGMAR;
+    case FLAG_LAPRAS_SPECIES_SWARMING:
+        return SPECIES_LAPRAS;
+    case FLAG_DRATINI_SPECIES_SWARMING:
+        return SPECIES_DRATINI;
+    case FLAG_SENTRET_SPECIES_SWARMING:
+        return SPECIES_SENTRET;
+    case FLAG_HOOTHOOT_SPECIES_SWARMING:
+        return SPECIES_HOOTHOOT;
+    case FLAG_LEDYBA_SPECIES_SWARMING:
+        return SPECIES_LEDYBA;
+    case FLAG_SPINARAK_SPECIES_SWARMING:
+        return SPECIES_SPINARAK;
+    case FLAG_MAREEP_SPECIES_SWARMING:
+        return SPECIES_MAREEP;
+    case FLAG_HOPPIP_SPECIES_SWARMING:
+        return SPECIES_HOPPIP;
+    case FLAG_MURKROW_SPECIES_SWARMING:
+        return SPECIES_MURKROW;
+    case SPECIES_MISDREAVUS:
+    case FLAG_MISDREAVUS_SPECIES_SWARMING:
+        return SPECIES_MISDREAVUS;
+    case FLAG_DUNSPARCE_SPECIES_SWARMING:
+        return SPECIES_DUNSPARCE;
+    case FLAG_GLIGAR_SPECIES_SWARMING:
+        return SPECIES_GLIGAR;
+    case FLAG_SNEASEL_SPECIES_SWARMING:
+        return SPECIES_SNEASEL;
+    case FLAG_REMORAID_SPECIES_SWARMING:
+        return SPECIES_REMORAID;
+    case FLAG_HOUNDOUR_SPECIES_SWARMING:
+        return SPECIES_HOUNDOUR;
+    case FLAG_MANTINE_SPECIES_SWARMING:
+        return SPECIES_MANTINE;
+    default:
+        return -1;
+    }
+}
+
 void ActivateNewSwarm()
 {
     u16 inactiveSwarms[NUM_SWARMS_FLAGS], inactive = 0;
@@ -230,4 +337,20 @@ bool8 HasSpeciesSwarmHappened(u16 species)
 {
     u16 flag = GetSwarmSpeciesFlag(species);
     return flag == -1 ? FALSE : FlagGet(flag);
+}
+
+u16 GetNextUnshowSpecies()
+{
+    for (u8 i = 0; i < NUM_SWARMS_FLAGS; i++)
+    {
+        u16 species = GetSpeciesFromSwarmFlag(FLAG_SWARMS_START + i);
+        if (!HasSpeciesSwarmHappened(species))
+            continue;
+        
+        if (!GetSetPokedexFlag(species, FLAG_GET_SEEN))
+            return species;
+        else if (species == SPECIES_NIDORAN_F && !GetSetPokedexFlag(SPECIES_NIDORAN_M, FLAG_GET_SEEN))
+            return SPECIES_NIDORAN_M;
+    }
+    return SPECIES_NONE;    
 }