+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## In-development
 - Modified the in-game trades to be more useful
-- Implemented the Swarm mechanic (Including swarm Pokemon showing up in the Pokedex correctly)
+- Implemented the Swarm mechanic:
+    - Swarm Pokemon show up in the Pokedex correctly (Need to fix bug showing Altering Cave Pokemon even when they're not there)
+    - Implemented Birch's book for active swarms
 - Increased the variety of wild Pokemon in some areas
 
 ## 0.1.0